refactor : 계층형 구조 > 도메인 구조로 프로젝트 구조 리팩토링
