+++ src/main/java/com/planner/server/domain/record/Record.java
@@ -1,5 +1,6 @@
-package com.planner.server.domain.entity;
+package com.planner.server.domain.record;
 
+import com.planner.server.domain.user.User;
 import com.sun.istack.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -14,18 +16,21 @@ import java.util.UUID;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class Attendance {
+public class Record {
 
     @Id
     @NotNull
     @Column(columnDefinition = "BINARY(16)")
     private UUID id;
 
+    private LocalDateTime recordStartTime;
+
+    private LocalDateTime recordEndTime;
+
+    private Duration totalStudyTime;
+
     @Setter
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
-    private User user;
-
-    private LocalDateTime createdAt;
-
+    User user;
 }
