+++ src/main/java/com/planner/server/domain/study_goal/StudyGoal.java
@@ -1,5 +1,6 @@
-package com.planner.server.domain.entity;
+package com.planner.server.domain.study_goal;
 
+import com.planner.server.domain.user.User;
 import com.sun.istack.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,27 +8,34 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.time.Duration;
+import java.time.LocalDate;
 import java.util.UUID;
 
 @Entity
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class RoomUser {
+public class StudyGoal {
 
     @Id
     @NotNull
     @Column(columnDefinition = "BINARY(16)")
     private UUID id;
 
+    private String goalTitle;
+
+    private String goalDetail;
+
+    private Duration goalTime;
+
+    private LocalDate startDate;
+
+    private LocalDate endDate;
+
     @Setter
     @ManyToOne
     @JoinColumn(name = "user_id")
     private User user;
-
-    @Setter
-    @ManyToOne
-    @JoinColumn(name = "study_room_id")
-    private StudyRoom studyRoom;
 
 }
